+++ app/Http/Controllers/Admin/Controller.php
@@ -1,0 +1,14 @@
+<?php
+
+/*
+ * Copyright © 2019 Dxvn, Inc. All rights reserved.
+ */
+
+namespace App\Http\Controllers\Admin;
+
+class Controller extends \App\Http\Controllers\Controller
+{
+    function __construct()
+    {
+    }
+}