+++ app/Http/Controllers/Admin/HomeController.php
@@ -1,17 +1,22 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
 use Illuminate\Http\Request;
 
-class Admin extends Controller
+class HomeController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        //
+
+        return view('admin/home', [
+            'data' => [
+                'name' => "home"
+            ]
+        ]);
     }
 
     /**