+++ resources/views/admin/home.blade.php
@@ -1,1 +1,4 @@
-admin
+@extends('admin.layout')
+@section('content')
+    {{ $data['name'] }}
+@endsection