+++ resources/views/layouts/app.blade.php
@@ -1,0 +1,40 @@
+<!DOCTYPE html>
+<html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
+
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="robots" content="noindex">
+
+    <!-- CSRF Token -->
+    <meta name="csrf-token" content="{{ csrf_token() }}">
+
+    <title>{{ config('app.name', 'diepxuan.com') }}</title>
+
+    <!-- Scripts -->
+    <script src="{{ asset('js/app.js') . '?v=' . uniqid() }}" defer></script>
+
+    <link rel="icon" href="{{ asset('favicon.ico') }}" type="image/x-icon" />
+    <link rel="shortcut icon" href="{{ asset('favicon.ico') }}" type="image/x-icon" />
+    <link rel="icon" type="image/png" href="{{ asset('favicon.png') }}" />
+
+    <!-- Fonts -->
+    <link rel="dns-prefetch" href="//fonts.gstatic.com">
+    {{-- <link href="https://fonts.googleapis.com/css?family=Arimo&family=Poppins" rel="stylesheet" type="text/css"> --}}
+
+    <!-- Styles -->
+    <link href="{{ asset('css/app.css') . '?v=' . uniqid() }}" rel="stylesheet">
+</head>
+
+<body class="d-none">
+    <div id="app">
+
+        <main class="container-fluid">
+            @yield('content')
+        </main>
+
+    </div>
+
+</body>
+
+</html>